suppress distracting message in ./g on some platform
we use perl to determine the absolute path of argument thought to be
fiels. but sometime they are not file at all and that call 'fail'.
The code is meant to deal with such case, but
on some platform that produce a scary message on stderr in the
middle of the build log.
This hide these messages.

Change-Id: I52d43e0b26847ab091d76fd446a05c4d84836a77