prevent crash in malformed pivot table loaded from .xls
ScDPMember::GetItemData() unconditionally returned a reference to a
ScDPItemData pointer obtained through ScDPSource::GetItemDataById()
ScDPTableData::GetMemberById() ScDPCache::GetItemDataById() that can be
null for malformed entries. Changed ScDPMember::GetItemData() to return
a pointer instead and adapted callers to check for null.

3.5.x in ScDPSource::GetItemDataById() had a check for null pointer and
added an empty ScDPItemData element to the cache for this case and
returned the pointer to that entry (marked as todo). This is not the
case anymore.

Change-Id: I241c232d7182f5d58e8531af540e69b26ab4888a