fix STL conversion eeb41cdd213ba0c9f454b639b418d9c64e072d43:
It turns out that this SvxMSDffShapeInfos has some, shall we say,
non-obvious properties, including a split personality that changes
its nature from checking for pointer equality to not doing so;
symptoms of this include a failing sw_subsequent_ww8export test.
Try to cure this madness by splitting the personalities into distinct
types, a set and a multiset.

Change-Id: Ib87368564b8e1f61379703034c0965422c6dcd0e