LookupTree for fast autocompletion lookups (by Nico Weyand).
LookupTree is a tree structure for fast autocompletion lookups.
Additionally the tree structure stores word probabilities, so each
autocompletion request returns a result with highest probability.
LatinLookupTree is an implementation which was designed to be even
faster and more efficient latin text, however it works with any kind
of unicode strings.

The tree structure was coded by Nico Weyand, Unicode strings support
and conversion to Libreoffice code structure was done by me.

Change-Id: I6549ee45d0952407b8a070f30ed0598fcb420aa7