Revert "sorted_vector: turn Find parameter into template"
This reverts commit 94aa98d3114aed610d4412ae45c636074c3b1198.

It was a nice idea, but C++ is not yet ready for it; with the travesty
of parametric polymorphism in C++ the find_unique inside the definition
of find_unique actually refers to find_unique<Value, Compare>, so there
is no way to actually refer to template<Value, Compare> find_unique
inside its definition.  Thanks to Luboš Luňák for explaining
the problem to me.  Somehow this does work in GCC 4.7 even with
-std=c++98, likely by accident.