remove now pointless pItemDesruptList
This pItemDesruptList is pointless now. We're registering the delete-when-idle
items in their ctor's and deregistering in their dtor's which get called from
the idle callback. The idea of pItemDesruptList appears to be for use in
deleting any items at exit time whose idle callback didn't get called already.

Which was done by DeleteOnIdleItems which itself was removed a while ago
because nothing was calling it (for years). Some experiments in making it a
simple singleton and/or one that hooks off default component dispose or
XDesktop dispose shows that its crash city, so lets just remove the unused
stuff outright rather than trying to fight to get some 12+ year unused appendix
used.

Change-Id: Ie0256d6987cf89a2a12db297065af09674547b3e