changes to "fdo#53002 Fix for default currency settings."
There can be only one usedInCompatibleFormatCodes="true" and because of
the new symbol with identical ISO 4217 TRY we need also
legacOnly="true" on the old entry. Additionally, as the TL symbol was
already used with TRL also that one needed legacyOnly="true" to not end
up with a list of duplicated TL formats in the number formatter.

Change-Id: I8e3ec2f75828b8f7c6a2077ef2fd30fc5e7a5642