Now MasterPage text fields work, list items are detected correctly,  some effect work on text.
Date/Time, Footer and Page Number text fields are displayed correctly again. 

Fixed several bugs related to synchronization of the two iterations on both meta comment 
actions and text paragraphs/text portions enumerations. At present list items inside a table 
or an OLE object are not exported correctly.

Enabled support for animating text paragraph. Both entrance and exit effects work on text 
shapes, on the contrary there is not yet support for emphasis effects such as changing font 
color and font properties. At present it is possible to apply an effect to the whole paragraph 
only, no support is provided for by word or by character effect variants. Another weak point 
of current implementation is that if you try to select a fragment of a text shape and in your 
selection is included an animatable paragraph the selection is not correct.