n#775899 SwXText::convertToTextFrame fix for multi-paragraph tables
The problem was that when we only had a table inside the frame, the
DelFullPara() call removed the last paragraph of the first cell, and
returned no error. Instead, check if we had to remove paragraphs both
before and after, and if so, use the manual remove all the time.

Change-Id: I9a3591ce9a92a2aca7d2f65b62fdf641f54e4f05