Revert "installer: Use hashref for replace_all_ziplistvariables_in_rtffile"
This reverts commit ef54f42dd23d803c2fcb8fec0b05bca030e0b5c0.  It started to
replace occurrences of "${...}" unknown to the installer with empty strings,
instead of keeping them as-is.  This caused the "${ORIGIN}" at the start of the
value for URE_BOOTSTRAP to disappear from the soffice ini-file (cf.
gid_Brand_Profileitem_Soffice_UreBootstrap in scp2/source/ooo/common_brand.scp),
making soffice fail to start completely.

Change-Id: Ia7a34179b78677a703fc487401e011073d14a176