fdo#49901: SwView::ExecSearch: fix Replace All cursor:
Because the "fix" for i#8288 only set the point of the cursor and not
the mark, it can happen that after a Replace All weird things are
selected; fix that (and i#8288 properly) by using the SwCrsrShell stack,
and don't do it when searching in the selection.
This cannot be done in SwCrsrShell itself because ExecSearch sets the
cursor to the beginning/end of the document already.
(regression from cda75d6f358d438f7ed36ed7fa296ac23e3c4a04)

Change-Id: Ie9f8d7da2fa26000a0237c24f8742e47ffa29213