fdo#39468: Translated German comments in sc/source/ui/undo
Eike and I reviewed the patch at the same time. However, he was was
faster ;-) Consequently, with this commit I include the fixes for
my remarks previously done in Gerrit as well as for Eike's comments.

This is a follow-up patch of 6d3047e7f4bd9f0bd18c7e3e2423a2fa4832ef80.

Change-Id: I6ee721f15852e1023c8b3e4f0b580435ee52c0c6