remove ascii-art and blank lines in sc/source/ui/undo
This is a follow-up patch to 6d3047e7f4bd9f0bd18c7e3e2423a2fa4832ef80.

However, since the rebasing process is not done yet, I submit it to
Gerrit, so that the changes are available as soon as they can be
applied again to the related files.

Change-Id: I61292e9d893839bd2f8b18a521a30e670f80f05b
Reviewed-on: https://gerrit.libreoffice.org/484
Reviewed-by: Kohei Yoshida <kohei.yoshida@gmail.com>
Tested-by: Kohei Yoshida <kohei.yoshida@gmail.com>