Basic IDE: Filling space of non-docking windows
Now if we make e.g. the stack window floating, then the watch window
will occupy the whole width of the line (and vice versa, also for the
Dialog Editor with object catalog and property browser). If we put it
back again, than we get back the previous state of the strip.

Change-Id: I634614ff2e9d5f790ad759656a79a38d49c4afc4