fdo#49350 Speedup entry painting for SvTreeListBox
Painting code for entry in SvTreeListBox switches between default
and highlight font (which is only the default font with a different
colour). This is done using SetFont which assumes that the font
changes entry dimensions, which then recalculates everything. This
is done for every entry so it is slow if there are a moderate number
of entries. Fix here is to call Control::SetFont - we can do this
because we know that switching between highlight and default font
wont change the font dimensions.

On the other hand - this highlighting entries when selected is
(at least on linux) already supported by the native widget.

Change-Id: I12d32f6f7b43e669442e462ad4dd80cde24b9e74