fix build of dictionaries with l10n enabled
It seems that

foo : BAR := xyz
foo $$(BAR)

does not really work, i.e., it does not generate a dep on xyz, but
$$(BAR) expands into an empty string. It is unclear to me whether it is
a bug in make or it does not work by design. Info documentation for make
only says that target-specific variable _assignments_ are evaluated in
the context of the target, so previously defined target-specific
variables shall be in scope. There is no mention whether this should
work for use of such a variable too (like in the example above). I have
always expected it does; experimental evidence shows it does not .-)

Change-Id: If99bdcacea403aa925127dd993a7e94ed7d57ae4