rhbz#836827: SQLFeatureNotSupportedException only in Java 6
...so if the patched hsqldb.jar is compiled with Java 7, it creates instances of
SQLFeatureNotSupportedException, and when at LO runtime a Java 5 gcj is used,
loading the relevant hsqldb.jar classes will fail with a NoClassDefFoundError
caused by a ClassNotFoundException.

Adding java.sql.Driver.getParentLogger in Java 7 was an incompatible change, but
it appears that the java.sql framework is prepared to handle old implementations
(for which the JVM will throw AbstractMethodError), so it appears that it works
to fake that AbstractMethodError here.

Change-Id: Id1fddc9cbdd8c88cd68fb5877ec59fe8d61e1562