rhbz#820283: fix crashes in DOCX table import:
It is apparently possible that writerfilter creates a table with
irregular structure, where the lines have varying numbers of boxes in
them.  This may cause crashes on later editing operations, e.g.
when SwTable::NewInsertCol() is unable to find boxes covering a
particular column.  To prevent this, add missing boxes in
SwNodes::TextToTable().

Change-Id: I423821645baeaf55595e4933e1ab8fb89b2099f3