n#782061 DOCX import: w:position should respect w:sz in w:rPrDefault
Subscript relative font size (w:position) is in percents in Writer, but
is in half points in Word. When checking for the default font size to
count the percent value, the mapper should search font size (w:sz) not
only in direct character properties, but also in default character
properties (w:rPrDefault).

Change-Id: I9286d44c6498c765ddfee795d50921ef58b80071