fdo#55496: fix lifecycle of SwMailMergeOutputPage:
Its member m_pDocumentPrinterCopy contains a SfxItemSet that points at
the SwDoc's item pool; the SwDoc is freed in
SwMailMergeWizardExecutor::LinkStubCancelHdl.  But the SwDoc::setPrinter
method does seem to work fine if the printer that is set is the same as
the one it already has, so it does not seem to be necessary to clone the
printer; just get it and set it again.

Change-Id: Ic591f68f6e0bab4e42cb22a9cb0a47e97f5e938f